third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-with-nested-shadow-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Test for http://crbug.com/360679, ::content rules in nested ShadowDOM should match correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(sandbox.querySelector('#host > .special')).backgroundColor is not "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE

I should not be red.
